Fixed SHA peripheral
Before, the SHA peripheral did not reset unless the "reset" register (or what we think is the reset register) at offset 0x0004 was written. However, at least the iPod Classic implementation does not make use of this, and instead writes 0x2 to the config register when starting a new SHA calculation.

We now use this to know when to start over and clear the internal buffer.

Furthermore, some leftovers from the iPod Touch port that did not apply to the S5L8702 were removed.